Split the sector-coupling output module into Hydrogen and Heat
  Per Andres' review on PR #128: openTEPES_OutputResultsSectorCoupling.py
  bundled two unrelated carriers. Move each writer into its own module --
  openTEPES_OutputResultsHydrogen.py (NetworkH2OperationResults) and
  openTEPES_OutputResultsHeat.py (NetworkHeatOperationResults). Pure move:
  both function bodies are unchanged. Import sites in openTEPES.py and
  __init__.py and the CHANGELOG are updated; OUTPUT_REGISTRY dispatch is
  unchanged.

  Verified bit-identical output (PYTHONHASHSEED=0) on 9n, 9n_heat, and a
  new 9n_H2 case, before vs after the split.
