+++ openTEPES/cases/9n_H2/oT_Data_Emission_9n_H2.csv
@@ -1,0 +1,2 @@
+﻿Period,Area,CO2Emission
+2030,Area1,
