+++ openTEPES/cases/9n_H2/oT_Data_Generation_9n_H2.csv
@@ -1,0 +1,19 @@
+Generator,Node,Technology,MutuallyExclusive,StorageType,OutflowsType,EnergyType,MustRun,OutflowsIncompatibility,NoOperatingReserve,BinaryInvestment,BinaryRetirement,BinaryCommitment,InitialPeriod,FinalPeriod,MaximumPower,MinimumPower,MaximumPowerHeat,MinimumPowerHeat,MaximumCharge,MinimumCharge,InitialStorage,MaximumStorage,MinimumStorage,Efficiency,ShiftTime,EFOR,RampUp,RampDown,UpTime,DownTime,StableTime,FuelCost,LinearTerm,ConstantTerm,OMVariableCost,OperReserveCost,StartUpCost,ShutDownCost,CO2EmissionRate,Availability,FixedInvestmentCost,FixedRetirementCost,FixedChargeRate,StorageInvestment,Inertia,MaximumReactivePower,MinimumReactivePower,InvestmentLo,InvestmentUp,RetirementLo,RetirementUp,ProductionFunctionHydro,ProductionFunctionH2,ProductionFunctionHeat,ProductionFunctionH2ToHeat
+Nuclear,Node_1,Nuclear,,,,,,,,,,,2020,2050,771.6,771.6,,,,,,,,,,,300,300,10.7,12,,1,15,,,,,,,1,,,,,,,,,,,,,,,
+DomesticCoal_Anthracite,Node_2,Coal,,,,,,,,,,,2020,2050,588,235.2,,,,,,,,,,,350,350,2,2,,0.02,2400,50000,6,,2,,,1,,,,,,,,,,,,,,,
+BrownLignite,Node_3,Coal,,,,,,,,,,,2020,2050,203.1,81.2,,,,,,,,,,,120,120,,,,0.02,2300,50000,6,,2,,,1,,,,,,,,,,,,,,,
+ImportedCoal_SubBituminous,Node_4,Coal,,,,,,,,,,,2020,2050,150.4,60.2,,,,,,,,,,,120,120,,,,0.02,2300,50000,6,,2,,,1,,,,,,,,,,,,,,,
+ImportedCoal_Bituminous,Node_1,Coal,,,,,,,,,,,2020,2050,194.4,77.8,,,,,,,,,,,120,120,,,,0.02,2200,50000,6,,2,,,1,,,,,,,,,,,,,,,
+CCGT_1,Node_2,Gas,,,,,,,,,,,2020,2050,500,100,,,,,,,,,,,500,500,,,,0.03,800,300000,6,,1,,,1,,,,,,,,,,,,,,,
+CCGT_2,Node_3,Gas,,,,,,,,,,,2020,2050,500,100,,,,,,,,,,,500,500,,,,0.03,900,300000,4,,1,,,1,,,,,,,,,,,,,,,
+CCGT_3,Node_4,Gas,,,,,,,,,,,2020,2050,500,100,,,,,,,,,,,500,500,,,,0.03,1000,300000,4,,1,,,1,,,,,,,,,,,,,,,
+CCGT_4,Node_1,Gas,,,,,,,,,,,2020,2050,667.5,133.5,,,,,,,,,,,667.5,667.5,,,,0.03,800,300000,4,,1,,,1,,,,,,,,,,,,,,,
+OCGT_1,Node_2,Gas,,,,,,,,,,,2020,2050,400,,,,,,,,,,,,400,400,,,,0.03,2000,,4,,0,,,1,,,,,,,,,,,,,,,
+OCGT_2,Node_3,Gas,,,,,,,,,,,2020,2050,400,,,,,,,,,,,,400,400,,,,0.03,2100,,4,,0,,,1,,,,,,,,,,,,,,,
+OCGT_3,Node_4,Gas,,,,,,,,,,,2020,2050,400,,,,,,,,,,,,400,400,,,,0.03,2200,,4,,0,,,1,,,,,,,,,,,,,,,
+FuelOilGas,Node_1,Oil,,,,,,,,,,,2020,2050,441.8,,,,,,,,,,,,441.8,441.8,,,,0.06,2000,,3,,1,,,1,,,,,,,,,,,,,,,
+WindFarm_1,Node_4,RES,,,,,,,,,,,2020,2050,450,,,,,,,,,,,,,,,,,,,,,,,,,0.15,,,,,,,,,,,,,,,
+SolarPV_1,Node_4,RES,,,,,,,,,,,2020,2050,100,,,,,,,,,,,,,,,,,,,,,,,,,0.000001,,,,,,,,,,,,,,,
+ESS1,Node_1,ESS,,Daily,Daily,,,,,,,,2020,2050,200,,,,200,,500,1000,,0.9,4,,,,,,,,,,,,,,,0.7,,,,,,,,,,,,,,,
+Electrolyzer_1,Node_1,Electrolyzer,,,,,,,,,,,2020,2050,,,,,200,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,60,,
+Electrolyzer_2,Node_4,Electrolyzer,,,,,,,,,,,2020,2050,,,,,200,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,60,,
