+++ openTEPES/cases/9n_H2/oT_Data_NetworkHydrogen_9n_H2.csv
@@ -1,0 +1,14 @@
+InitialNode,FinalNode,Circuit,InitialPeriod,FinalPeriod,Length,TTC,TTCBck,SecurityFactor,FixedInvestmentCost,FixedChargeRate,BinaryInvestment,InvestmentLo,InvestmentUp
+Node_1,Node_6,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_2,Node_3,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_2,Node_6,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_3,Node_4,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_3,Node_6,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_4,Node_5,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_4,Node_6,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_4,Node_9,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_6,Node_7,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_6,Node_8,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_7,Node_8,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_8,Node_9,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
+Node_1,Node_4,h2pipe1,2020,2050,100,2.0,,1.0,,,,,
