+++ openTEPES/cases/9n_H2/oT_Data_NodeLocation_9n_H2.csv
@@ -1,0 +1,10 @@
+Node,Latitude,Longitude
+Node_1,43.36666667,-8.383333333
+Node_2,43.53333333,-5.7
+Node_3,43.26222222,-2.953333333
+Node_4,40.41888889,-3.691944444
+Node_5,41.3825,2.176944444
+Node_6,37.38333333,-5.983333333
+Node_7,36.51666667,-6.283333333
+Node_8,39.46666667,-0.375
+Node_9,36.71666667,-4.416666667
