+++ openTEPES/cases/9n_H2/oT_Data_Option_9n_H2.csv
@@ -1,0 +1,2 @@
+IndBinGenInvest,IndBinGenRetirement,IndBinRsrInvest,IndBinNetInvest,IndBinNetH2Invest,IndBinNetHeatInvest,IndBinGenOperat,IndBinNetLosses,IndBinLineCommit,IndBinSingleNode,IndBinGenRamps,IndBinGenMinTime
+0,0,0,0,0,0,0,1,0,0,0,0
