+++ openTEPES/cases/9n_H2/oT_Data_Parameter_9n_H2.csv
@@ -1,0 +1,2 @@
+ENSCost,HNSCost,HTNSCost,CO2Cost,UpReserveActivation,DwReserveActivation,MinRatioDwUp,MaxRatioDwUp,SBase,ReferenceNode,TimeStep,EconomicBaseYear,AnnualDiscountRate
+10000,10000,10000,25,0.25,0.3,0,1,100,Node_4,2,2020,0
