+++ openTEPES/cases/9n_H2/oT_Data_Period_9n_H2.csv
@@ -1,0 +1,2 @@
+﻿Period,Weight
+2030,1
