+++ openTEPES/cases/9n_H2/oT_Data_RESEnergy_9n_H2.csv
@@ -1,0 +1,2 @@
+Period,Area,RESEnergy
+2030,Area1,1000
