+++ openTEPES/cases/9n_H2/oT_Data_ReserveMargin_9n_H2.csv
@@ -1,0 +1,2 @@
+﻿Period,Area,ReserveMargin
+2030,Area1,1.1
