+++ openTEPES/cases/9n_H2/oT_Dict_Generation_9n_H2.csv
@@ -1,0 +1,19 @@
+Generator
+Nuclear
+DomesticCoal_Anthracite
+BrownLignite
+ImportedCoal_SubBituminous
+ImportedCoal_Bituminous
+CCGT_1
+CCGT_2
+CCGT_3
+CCGT_4
+OCGT_1
+OCGT_2
+OCGT_3
+FuelOilGas
+WindFarm_1
+SolarPV_1
+ESS1
+Electrolyzer_1
+Electrolyzer_2
