+++ openTEPES/cases/9n_H2/oT_Dict_NodeToZone_9n_H2.csv
@@ -1,0 +1,10 @@
+Node,Zone
+Node_1,Zone1
+Node_2,Zone2
+Node_3,Zone3
+Node_4,Zone4
+Node_5,Zone5
+Node_6,Zone6
+Node_7,Zone7
+Node_8,Zone8
+Node_9,Zone9
