+++ openTEPES/cases/9n_H2/oT_Dict_Region_9n_H2.csv
@@ -1,0 +1,31 @@
+Region
+AT
+BE
+BG
+CH
+CY
+CZ
+DE
+DK
+EE
+EL
+ES
+FI
+FR
+HR
+HU
+IE
+IT
+LT
+LU
+LV
+MT
+NL
+NO
+PL
+PT
+RO
+SE
+SI
+SK
+UK
