+++ openTEPES/cases/9n_H2/oT_Dict_Storage_9n_H2.csv
@@ -1,0 +1,3 @@
+StorageType
+Daily
+Weekly
