+++ openTEPES/cases/9n_H2/oT_Dict_Technology_9n_H2.csv
@@ -1,0 +1,8 @@
+Technology
+Nuclear
+Coal
+Gas
+Oil
+RES
+ESS
+Electrolyzer
