+++ openTEPES/cases/9n_H2/oT_Dict_ZoneToArea_9n_H2.csv
@@ -1,0 +1,10 @@
+Zone,Area
+Zone1,Area1
+Zone2,Area1
+Zone3,Area1
+Zone4,Area1
+Zone5,Area1
+Zone6,Area1
+Zone7,Area1
+Zone8,Area1
+Zone9,Area1
