+++ openTEPES/cases/9n_H2/oT_Dict_Zone_9n_H2.csv
@@ -1,0 +1,10 @@
+Zone
+Zone1
+Zone2
+Zone3
+Zone4
+Zone5
+Zone6
+Zone7
+Zone8
+Zone9
